Fragssss

+++ src/components/PulsatingBall/frag.js
@@ -1,11 +1,11 @@
 export default `
 // set the precision of floating point numbers
-precision highp float;
+precision mediump float;
 
 // this value is populated by the vertex shader
 varying vec3 fragColor;
 
 void main() {
   // gl_FragColor is a special variable that holds the color of a pixel
-  gl_FragColor = vec4(fragColor, 1);
+  gl_FragColor = vec4(fragColor, 1.0);
 }`
